+++ coroutines-interop/src/commonMain/kotlin/com/badoo/reaktive/coroutinesinterop/ObservableAsFlow.kt
@@ -4,7 +4,6 @@ import com.badoo.reaktive.disposable.Disposable
 import com.badoo.reaktive.disposable.DisposableWrapper
 import com.badoo.reaktive.observable.Observable
 import com.badoo.reaktive.observable.ObservableObserver
-import com.badoo.reaktive.utils.ensureNeverFrozen
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
@@ -13,8 +12,6 @@ import kotlinx.coroutines.flow.channelFlow
 @ExperimentalCoroutinesApi
 fun <T> Observable<T>.asFlow(): Flow<T> =
     channelFlow {
-        channel.ensureNeverFrozen()
-
         val disposableWrapper = DisposableWrapper()
 
         val observer =
