+++ coroutines-interop/src/commonTest/kotlin/com/badoo/reaktive/coroutinesinterop/MaybeFromCoroutineTest.kt
@@ -3,7 +3,13 @@ package com.badoo.reaktive.coroutinesinterop
 import com.badoo.reaktive.test.base.assertError
 import com.badoo.reaktive.test.maybe.assertSuccess
 import com.badoo.reaktive.test.maybe.test
+import com.badoo.reaktive.utils.atomic.AtomicReference
+import com.badoo.reaktive.utils.atomic.getValue
+import com.badoo.reaktive.utils.atomic.setValue
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.isActive
 import kotlin.test.Test
+import kotlin.test.assertFalse
 
 class MaybeFromCoroutineTest {
 
@@ -28,5 +34,15 @@ class MaybeFromCoroutineTest {
         val observer = maybeFromCoroutine { throw error }.test()
 
         observer.assertError(error)
+    }
+
+    @Test
+    fun cancels_coroutine_WHEN_disposable_is_disposed() {
+        var scope by AtomicReference<CoroutineScope?>(null)
+        val observer = maybeFromCoroutine { scope = this }.test()
+
+        observer.dispose()
+
+        assertFalse(scope!!.isActive)
     }
 }
