--- coroutines-interop/src/jvmJsCommonTest/kotlin/com/badoo/reaktive/coroutinesinterop/CompletableFromCoroutineJvmJsTest.kt
@@ -1,28 +1,0 @@
-package com.badoo.reaktive.coroutinesinterop
-
-import com.badoo.reaktive.test.completable.test
-import kotlinx.coroutines.isActive
-import kotlin.coroutines.Continuation
-import kotlin.coroutines.suspendCoroutine
-import kotlin.test.Test
-import kotlin.test.assertFalse
-
-class CompletableFromCoroutineJvmJsTest {
-
-    @Test
-    fun cancels_coroutine_WHEN_disposable_is_disposed() {
-        lateinit var continuation: Continuation<Nothing>
-
-        val observer =
-            completableFromCoroutine {
-                suspendCoroutine<Unit> {
-                    continuation = it
-                }
-            }
-                .test()
-
-        observer.dispose()
-
-        assertFalse(continuation.context.isActive)
-    }
-}
