--- coroutines-interop/src/nativeCommonTest/kotlin/com/badoo/reaktive/coroutinesinterop/test/CoroutineCancellationVerifierExt.kt
@@ -1,9 +1,0 @@
-package com.badoo.reaktive.coroutinesinterop.test
-
-import com.badoo.reaktive.disposable.Disposable
-
-fun CoroutineCancellationVerifier.verifyCancellation(disposable: Disposable) {
-    awaitSuspension()
-    disposable.dispose()
-    awaitCancellation()
-}
