+++ includedBuild/dependencies/src/main/kotlin/com/badoo/reaktive/dependencies/Deps.kt
@@ -3,7 +3,7 @@ package com.badoo.reaktive.dependencies
 object Deps {
 
     private const val kotlinVersion = "1.4.0"
-    private const val coroutinesVersion = "1.3.9"
+    private const val coroutinesVersion = "1.3.9-native-mt"
     private const val detektVersion = "1.9.1"
     private const val asmVersion = "6.0"
